+++ brainpy/core/network.py
@@ -1,18 +1,16 @@
 # -*- coding: utf-8 -*-
 
 import time
+from collections import OrderedDict
 
 import numpy as np
 from numba import cuda
 
-from .base import Ensemble
-from .constants import INPUT_OPERATIONS
-from .constants import SCALAR_MODE
-from .neurons import NeuGroup
-from .synapses import SynConn
-from .. import profile
-from .. import tools
+from . import base
+from . import constants
+from . import neurons
 from . import utils
+from .. import profile
 from ..errors import ModelUseError
 
 __all__ = [
@@ -28,55 +26,42 @@ class Network(object):
     actually runs the simulation. The main loop runs according to user add
     orders. The objects in the `Network` are accessible via their names, e.g.
     `net.name` would return the `object` (including neurons and synapses).
-
     """
 
-    def __init__(self, *args, mode='once', **kwargs):
-        # store and neurons and synapses
-        self._all_neu_groups = []
-        self._all_syn_conns = []
-        self._all_objects = []
-
-        # store all objects
-        self._objsets = dict()
-
+    def __init__(self, *args, mode=None, **kwargs):
         # record the current step
         self.t_start = 0.
         self.t_end = 0.
-        self.t_duration = 0.
 
-        # add objects
+        # store all objects
+        self._all_objects = OrderedDict()
         self.add(*args, **kwargs)
 
-        # check
-        assert mode in ['once', 'repeat']
-        self.mode = mode
-
+        # store the step function
         self._step_func = None
 
+        if isinstance(mode, str):
+            print('The "repeat" mode of the network is set to the default. '
+                  'After version 0.4.0, "mode" setting will be removed.')
+
     def _add_obj(self, obj, name=None):
-        # check object type
-        self._all_objects.append(obj)
-        if isinstance(obj, NeuGroup):
-            self._all_neu_groups.append(obj)
-        elif isinstance(obj, SynConn):
-            self._all_syn_conns.append(obj)
-        else:
-            raise ValueError(f'Unknown object type "{type(obj)}". Network only support NeuGroup and SynConn.')
-
-        # check object name
+        # 1. check object type
+        if not isinstance(obj, neurons.Ensemble):
+            raise ValueError(f'Unknown object type "{type(obj)}". Network '
+                             f'only supports NeuGroup and SynConn.')
+        # 2. check object name
         name = obj.name if name is None else name
-        if name in self._objsets:
-            raise KeyError(f'Name "{name}" has been used in the network, please change another name.')
-
-        # add object in the network
-        self._objsets[name] = obj
+        if name in self._all_objects:
+            raise KeyError(f'Name "{name}" has been used in the network, '
+                           f'please change another name.')
+        self._all_objects[name] = obj
+        # 3. add object to the network
         setattr(self, name, obj)
         if obj.name != name:
             setattr(self, obj.name, obj)
 
     def add(self, *args, **kwargs):
-        """Add object (neurons or synapses or monitor) to the network.
+        """Add object (neurons or synapses) to the network.
 
         Parameters
         ----------
@@ -92,15 +77,28 @@ class Network(object):
         for name, obj in kwargs.items():
             self._add_obj(obj, name)
 
-    def _format_inputs(self, inputs, run_length):
-        # check
-        try:
-            assert isinstance(inputs, (tuple, list))
-        except AssertionError:
-            raise ModelUseError('"inputs" must be a tuple/list.')
+    def format_inputs(self, inputs, run_length):
+        """Format the user defined inputs.
 
+        Parameters
+        ----------
+        inputs : tuple
+            The inputs.
+        run_length : int
+            The running length.
+
+        Returns
+        -------
+        formatted_input : dict
+            The formatted input.
+        """
+
+        # 1. format the inputs to standard
+        #    formats and check the inputs
+        if not isinstance(inputs, (tuple, list)):
+            raise ModelUseError('"inputs" must be a tuple/list.')
         if len(inputs) > 0 and not isinstance(inputs[0], (list, tuple)):
-            if isinstance(inputs[0], Ensemble):
+            if isinstance(inputs[0], base.Ensemble):
                 inputs = [inputs]
             else:
                 raise ModelUseError('Unknown input structure. Only supports "(target, key, value, [operation])".')
@@ -108,17 +106,18 @@ class Network(object):
             if not 3 <= len(inp) <= 4:
                 raise ModelUseError('For each target, you must specify "(target, key, value, [operation])".')
             if len(inp) == 4:
-                if inp[3] not in INPUT_OPERATIONS:
+                if inp[3] not in constants.INPUT_OPERATIONS:
                     raise ModelUseError(f'Input operation only support '
-                                        f'"{list(INPUT_OPERATIONS.keys())}", not "{inp[3]}".')
+                                        f'"{list(constants.INPUT_OPERATIONS.keys())}", '
+                                        f'not "{inp[3]}".')
 
-        # format input
+        # 2. format inputs
         formatted_inputs = {}
         for inp in inputs:
             # target
             if isinstance(inp[0], str):
                 target = getattr(self, inp[0]).name
-            elif isinstance(inp[0], Ensemble):
+            elif isinstance(inp[0], base.Ensemble):
                 target = inp[0].name
             else:
                 raise KeyError(f'Unknown input target: {str(inp[0])}')
@@ -149,38 +148,55 @@ class Network(object):
             else:
                 ops = '+'
 
-            # append
+            # final result
             if target not in formatted_inputs:
                 formatted_inputs[target] = []
             format_inp = (key, val, ops, data_type)
             formatted_inputs[target].append(format_inp)
-
         return formatted_inputs
 
     def build(self, run_length, inputs=()):
-        assert isinstance(run_length, int)
-        code_scopes = {}
-        code_lines = ['# network step function\n'
-                      'def step_func(_t, _i, _dt):']
+        """Build the network.
+
+        Parameters
+        ----------
+        run_length : int
+            The running length.
+        inputs : tuple, list
+            The user-defined inputs.
+
+        Returns
+        -------
+        step_func : callable
+            The step function.
+        """
+        if not isinstance(run_length, int):
+            raise ModelUseError(f'The running length must be an int, but we get {run_length}')
 
         # inputs
-        format_inputs = self._format_inputs(inputs, run_length)
+        format_inputs = self.format_inputs(inputs, run_length)
 
-        for obj in self._all_objects:
+        # codes for step function
+        code_scopes = {}
+        code_lines = ['# network step function\ndef step_func(_t, _i, _dt):']
+        for obj in self._all_objects.values():
             if profile.run_on_gpu():
-                if obj.model.mode != SCALAR_MODE:
-                    raise ModelUseError('GPU mode only support scalar-based mode.')
-
+                if obj.model.mode != constants.SCALAR_MODE:
+                    raise ModelUseError(f'GPU mode only support scalar-based mode. '
+                                        f'But {obj.model} is a {obj.model.mode}-based model.')
             code_scopes[obj.name] = obj
             code_scopes[f'{obj.name}_runner'] = obj.runner
-            lines_of_call = obj._build(inputs=format_inputs.get(obj.name, None), mon_length=run_length)
+            lines_of_call = obj.build(inputs=format_inputs.get(obj.name, None), mon_length=run_length)
             code_lines.extend(lines_of_call)
         if profile.run_on_gpu():
             code_scopes['cuda'] = cuda
         func_code = '\n  '.join(code_lines)
+
+        # compile the step function
         exec(compile(func_code, '', 'exec'), code_scopes)
         step_func = code_scopes['step_func']
 
+        # show
         if profile.show_format_code():
             utils.show_code_str(func_code.replace('def ', f'def network_'))
         if profile.show_code_scope():
@@ -188,7 +204,8 @@ class Network(object):
 
         return step_func
 
-    def run(self, duration, inputs=(), report=False, report_percent=0.1):
+    def run(self, duration, inputs=(), report=False, report_percent=0.1,
+            data_to_host=False):
         """Run the simulation for the given duration.
 
         This function provides the most convenient way to run the network.
@@ -204,6 +221,8 @@ class Network(object):
             Report the progress of the simulation.
         report_percent : float
             The speed to report simulation progress.
+        data_to_host : bool
+            Transfer the gpu data to cpu.
         """
         # check the duration
         # ------------------
@@ -220,22 +239,14 @@ class Network(object):
         ts = np.asarray(np.arange(start, end, dt), dtype=np.float_)
         run_length = ts.shape[0]
 
-        if self.mode != 'repeat' or self._step_func is None:
+        if self._step_func is None:
             # initialize the function
             # -----------------------
             self._step_func = self.build(run_length, inputs)
-            self.t_duration = end - start
         else:
-            # check running duration
-            # ----------------------
-            if self.t_duration != (end - start):
-                raise ModelUseError(f'Each run in "repeat" mode must be done '
-                                    f'with the same duration, but got '
-                                    f'{self.t_duration} != {end - start}.')
-
             # check and reset inputs
             # ----------------------
-            formatted_inputs = self._format_inputs(inputs, run_length)
+            formatted_inputs = self.format_inputs(inputs, run_length)
             for obj_name, inps in formatted_inputs.items():
                 obj = getattr(self, obj_name)
                 obj_inputs = obj.runner._inputs
@@ -255,6 +266,11 @@ class Network(object):
                     all_keys.remove(key)
                 if len(all_keys):
                     raise ModelUseError(f'The inputs of {all_keys} in {obj_name} are not provided.')
+
+            # reset monitors
+            # --------------
+            for obj in self._all_objects.values():
+                obj.reshape_mon(run_length)
 
         dt = self.dt
         if report:
@@ -281,10 +297,10 @@ class Network(object):
 
         # format monitor
         # --------------
-        for obj in self._all_objects:
-            if profile.run_on_gpu():
-                obj.runner.gpu_data_to_cpu()
+        for obj in self._all_objects.values():
             obj.mon['ts'] = self.ts
+            if data_to_host and profile.run_on_gpu():
+                obj.runner.gpu_data_to_cpu()
 
     @property
     def ts(self):